(maint) Move BoltCompiler to a separate file
Start breaking Bolt::Catalog into smaller pieces by moving the
BoltCompiler to its own file.
